Supplies a makefile for development
Makefile can optionally accept WITHQT4 arugment which will build an old
version of QT4 which is an optional prereq for ETE. Assumes Debian-like Linux
will use apt to instlal system prereqs. Makes a virtual environment for
python3 and uses pip to install as a development. This should be the minimal
target for running nosetests