Adds the old cli, modifies setup to point to simplified entry
Copied the __main__.py script from the old repo, but it seemed it was doing too
much so I removed it and pointed it to a main() in cli.py. Removes one file
and two redirected functions calls. Loses a little bit of flexibility of the
__main__ idiom, ie python -m evoc, but as long as the package is installed
this should not be an issue